requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.3
Cython==3.1.1
dl==0.1.0
docutils==0.21.2
filterpy==1.4.5
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
lap==0.5.12
lockfile==0.12.2
mock==5.2.0
numpy==2.2.6
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.2.1
protobuf==6.31.0
pyOpenSSL==25.1.0
pywin32==310
PyYAML==6.0.2
PyYAML==6.0.2
railroad==0.5.0
scipy==1.15.3
Sphinx==8.2.3
thread==2.0.5
tornado==6.4.2
trove_classifiers==2025.5.9.12
truststore==0.10.1
ultralytics==8.3.144
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
